requirements.txt
# ============================================================
# Speech Enhancement Fine-Tuning - Requirements
# ============================================================

# Pre-trained Model
denoiser>=0.1.5          # Facebook Demucs speech enhancement

# Core Deep Learning
torch>=2.1.0
torchaudio>=2.1.0

# Audio Processing
soundfile>=0.12.1
pyloudnorm>=0.1.1
scipy>=1.11.0
numpy>=1.24.0

# Metrics
pesq>=0.0.4
pystoi>=0.3.3

# Training Utilities
tensorboard>=2.14.0
tqdm>=4.66.0

# Config
PyYAML>=6.0.1

# Streamlit App
streamlit>=1.30.0

# Data
datasets>=2.14.0
huggingface-hub>=0.19.0
